wip: merge develop - fix conflicts and relocate components
