+++ src/ui/pages/Connections/components/ConnectionsOptionModal/ConnectionsOptionModal.scss
