+++ src/ui/pages/Connections/components/IdentifierSelectorModal/IdentifierSelectorModal.scss
