+++ src/ui/pages/Connections/components/IdentifierSelectorModal/IdentifierSelectorModal.test.tsx
@@ -1,12 +1,11 @@
 import { setupIonicReact } from "@ionic/react";
 import { mockIonicReact, waitForIonicReact } from "@ionic/react-test-utils";
 import { fireEvent, render, waitFor } from "@testing-library/react";
-import { act } from "react-dom/test-utils";
+import { act, ReactNode } from "react";
 import { Provider } from "react-redux";
 import configureStore from "redux-mock-store";
-import { ReactNode } from "react";
-import EN_TRANSLATIONS from "../../../../../../../locales/en/en.json";
-import { TabsRoutePath } from "../../../../../../../routes/paths";
+import EN_TRANSLATIONS from "../../../../../locales/en/en.json";
+import { TabsRoutePath } from "../../../../../routes/paths";
 import { IdentifierSelectorModal } from "./IdentifierSelectorModal";
 
 setupIonicReact();
