+++ src/ui/pages/IncomingRequest/components/SignRequest.scss
