+++ src/ui/pages/WalletConnect/WalletConnect.scss
