+++ src/ui/pages/WalletConnect/WalletConnect.types.ts
