+++ src/ui/pages/WalletConnect/index.ts
